Fix eslintrc erros
