+++ package.json
@@ -16,15 +16,12 @@
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.2",
     "babel-preset-env": "^1.7.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^7.0.0",
     "css-loader": "^5.0.2",
     "dotenv-webpack": "^6.0.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
     "html-loader": "^2.0.0",
-    "html-webpack-plugin": "^5.1.0",
     "mini-css-extract-plugin": "^1.3.7",
     "node-sass": "^5.0.0",
     "sass-loader": "^11.0.1",
@@ -32,13 +29,16 @@
     "stylelint-config-standard": "^20.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.17.2",
-    "terser-webpack-plugin": "^5.1.1",
-    "webpack": "^5.22.0",
     "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^7.0.0",
+    "html-webpack-plugin": "^5.1.0",
+    "terser-webpack-plugin": "^5.1.1",
     "phaser": "^3.52.0",
+    "webpack": "^5.22.0",
     "webpack-merge": "^5.7.3"
   }
 }
